requirements.txt
Flask>=2.2
Werkzeug>=2.2.0
PyJWT==2.3.0
